# autoops_jenkins/docs/swagger.yaml
definitions:
...
  forms.LoginForm:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - username
    type: object
  forms.UserAddForm:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  forms.UserInfo:
    properties:
      password:
        maxLength: 20
        minLength: 3
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  forms.UsernameForm:
    properties:
      username:
        type: string
    required:
    - username
    type: object
info:
  contact: {}
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登陆信息
        in: body
        name: forms.LoginForm
        required: true
        schema:
          $ref: '#/definitions/forms.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 用户登陆
      tags:
      - 登陆认证
...
  /v1/user/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Insert your access Token
        in: header
        name: Token
        required: true
        type: string
      - description: 用户信息
        in: body
        name: forms.UserAddForm
        required: true
        schema:
          $ref: '#/definitions/forms.UserAddForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 用户添加
      tags:
      - 用户管理
  /v1/user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Insert your access Token
        in: header
        name: Token
        required: true
        type: string
      - description: 用户信息
        in: body
        name: forms.UsernameForm
        required: true
        schema:
          $ref: '#/definitions/forms.UsernameForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 用户删除
      tags:
      - 用户管理
  /v1/user/get:
    post:
      parameters:
      - description: Insert your access Token
        in: header
        name: Token
        required: true
        type: string
      - description: 用户信息
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/forms.UsernameForm'
      responses:
        "200":
          description: OK
      summary: 用户查询(单个)
      tags:
      - 用户管理
  /v1/user/getall:
    get:
      consumes:
      - application/json
      parameters:
      - description: Insert your access Token
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 用户查询(全部)
      tags:
      - 用户管理
  /v1/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Insert your access Token
        in: header
        name: Token
        required: true
        type: string
      - description: 用户信息
        in: body
        name: forms.UserInfo
        required: true
        schema:
          $ref: '#/definitions/forms.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 用户修改
      tags:
      - 用户管理
swagger: "2.0"